Move storage.rs to a directory module.
This splits history specific stuff into a history submodule, but keeps
stuff that needs to be shared between history and bookmarks in the top-level
storage module.
